Create MySQL Telemetry attribute
