--- module/ldbc-mysql/shared/src/main/scala/ldbc/mysql/telemetry/PoolMetricsState.scala
@@ -1,20 +1,0 @@
-/**
- * Copyright (c) 2023-2026 by Takahiko Tominaga
- * This software is licensed under the MIT License (MIT).
- * For more information see LICENSE or https://opensource.org/licenses/MIT
- */
-
-package ldbc.mysql.telemetry
-
-/**
- * Snapshot of pool state used by Observable BatchCallback to report metrics.
- *
- * @param idleCount the number of idle connections
- * @param usedCount the number of connections currently in use
- * @param pendingRequestCount the number of pending requests waiting for a connection
- */
-case class PoolMetricsState(
-  idleCount:           Long,
-  usedCount:           Long,
-  pendingRequestCount: Long
-)
